# release-server/openapi.yaml
openapi: 3.0.0
info:
  title: Spec Kit Release Server
  version: 0.1.0
  description: Lightweight release server for Spec Kit templates.
paths:
  /latest:
    get:
      summary: Get latest release metadata
      description: Returns a GitHub-compatible release object with assets.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Release'
  /upload:
    post:
      summary: Upload a package
      description: Upload a new package file.
      parameters:
        - name: overwrite
          in: query
          description: Overwrite valid file if it exists
          required: false
          schema:
            type: boolean
            default: false
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  minLength: 1
              required:
                - file
      responses:
        '200':
          description: Upload successful
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '409':
          description: File already exists (conflict)
        '422':
          description: Validation Error
  /assets/{filename}:
    get:
      summary: Download asset
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File stream
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: File not found
    delete:
      summary: Delete package
      description: Deletes a package file from storage.
      security:
        - bearerAuth: []
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Package deleted
        '401':
          description: Unauthorized
        '404':
          description: File not found
  /packages:
    get:
      summary: List packages
      description: Returns list of packages in HTML or JSON.
      parameters:
        - name: format
          in: query
          description: Response format (json or html). Overrides Accept header.
          required: false
          schema:
            type: string
            enum: [json, html]
      responses:
        '200':
          description: List of packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Package'
            text/html:
              schema:
                type: string
        '422':
          description: Validation Error
  /healthz:
    get:
      summary: Liveness check
      description: Returns 200 OK if server is alive.
      responses:
        '200':
          description: Server is alive
          content:
            text/plain:
              schema:
                type: string
                example: "OK"
  /readyz:
    get:
      summary: Readiness check
      description: Returns 200 OK if server is ready to accept traffic.
      responses:
        '200':
          description: Server is ready
          content:
            text/plain:
              schema:
                type: string
                example: "OK"
        '503':
          description: Server is not ready

components:
  schemas:
    Release:
      type: object
      properties:
        tag_name:
          type: string
          example: "latest"
        assets:
          type: array
          items:
            $ref: '#/components/schemas/Asset'
    Asset:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
        browser_download_url:
          type: string
    Package:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
        created_at:
          type: string
          format: date-time
        sha256:
          type: string
          nullable: true
          description: SHA256 checksum of the package file (hex digest). Null for legacy files.